dt-bindings: media: Add the Raspberry Pi HEVC decoder
Adds a binding for the HEVC decoder IP owned by Raspberry Pi.

Instantiations of the decoder IP can currently be found in
the Broadcom BCM2711 and BCM2712 SoCs.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

---

Updated based on the thread in v3 [1] with Krzystof to drop to only SoC
specific compatible strings.

This still feels wrong to me as this is NOT Broadcom IP - it is solely
owned by Raspberry Pi.

It is the same situation as for the Raspberry Pi ISP Backend where
it was proposed to have
  compatible = "brcm,pispbe-bcm2712", "raspberrypi,pispbe";
in [2]. That proposal was viewed as acceptable by both Rob[3] and Conor[4],
and the binding given a Reviewed-by by Rob[5].

Yes it is found as a block inside an SoC, but the same could be said
for a Synopsys I2C controller ("snps,designware-i2c"), Cadence ethernet
MAC ("cdns,macb"), ARM CPU (eg "arm,cortex-a53"), or any number of other
examples. They are all 3rd party IP blocks found in SoCs, have an SoC specific
compatible defined, but acknowledge the IP owner and use a common fallback
compatible where possible.

Please could we have confirmation that this really is what is required
by device tree?

Thanks
  Dave

[1] https://lore.kernel.org/linux-media/20250423-media-rpi-hevc-dec-v3-3-8fd3fad1d6fb@raspberrypi.com/
[2] https://lore.kernel.org/linux-media/20240213104405.GB5012@pendragon.ideasonboard.com/
[3] https://lore.kernel.org/linux-media/20240215140058.GA4147262-robh@kernel.org/
[4] https://lore.kernel.org/linux-media/20240213-landlady-backstab-6e7da824c99a@spud/
[5] https://lore.kernel.org/linux-media/170932821578.3062059.18150159256247255021.robh@kernel.org/