staging: vchiq_arm: Set up dma ranges on child devices
The VCHIQ driver now loads the audio, camera, codec, and vc-sm
drivers as platform drivers. However they were not being given
the correct DMA configuration.

Call of_dma_configure with the parent (VCHIQ) parameters to be
inherited by the child. Although it is no longer necessary for
vchiq's children to have a different DMA configuration to the
parent, they do still need explicitly to have their DMA
configuration set - to be that of the parent.

vchiq on Pi4 is no longer under the soc node, therefore it
doesn't get the dma-ranges for the VPU. Switch to using the
configuration of the old dma controller as that will set the
dma-ranges correctly.

vchiq kernel clients are now instantiated as platform drivers rather
than using DT, but the children of the vchiq interface may still
benefit from access to DT properties. Give them the option of a
a sub-node of the vchiq parent for configuration and to allow
them to be disabled.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Phil Elwell <phil@raspberrypi.com>